+++ packages/remote/src/protocol.ts
@@ -1,8 +1,10 @@
 /**
- * @overview Public third-party Protocol implementor entry point.
+ * @overview Public Protocol implementation and third-party implementor entry point.
  * @author AEPKILL
  * @created 2026-08-19 00:00:00
  */
+
+export { createRpcProtocol } from "@/factories/rpc-protocol.factory";
 
 export type {
 	IRpcApplicationArgumentsSnapshot,